mcp/gmail-mcp/pyproject.toml
[project]
name = "gmail-mcp"
version = "0.1.0"
description = "Multi-account Gmail MCP server with search and summarization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.116.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
]

[project.scripts]
gmail-mcp = "gmail_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gmail_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
